Moved out Filter code to AB2D-Filters repository. Still need a way to link all of it.
