+++ pom.xml
@@ -10,7 +10,6 @@
         <module>api</module>
         <module>worker</module>
         <module>bfd</module>
-        <module>filter</module>
         <module>audit</module>
         <module>hpms</module>
         <module>e2e-test</module>
